MONGOID-4030 create typed array class factory and add evolve method
