+++ lib/mongoid/extensions.rb
@@ -57,3 +57,4 @@ require "mongoid/extensions/time"
 require "mongoid/extensions/time_with_zone"
 require "mongoid/extensions/true_class"
 require "mongoid/extensions/typed_array"
+require "mongoid/extensions/typed_array_class_factory"
